Add floating type specification through cmake.

+++ benchmark/benchmark_metapod_simple_humanoid.cc
@@ -26,7 +26,7 @@
 #include "benchmark.hh"
 using namespace metapod::benchmark;
 
-typedef double LocalFloatType;
+typedef float LocalFloatType;
 int main()
 {
   benchmark<metapod::simple_humanoid<LocalFloatType> >::run();
